Create Normal_samples_seurat_analysis
This markdown file details how the objects used for analysis in Gonzalez & Schrode et al 2022 were generated, including subsetting of the cardiac regions of interest as well for each individual sample (cardiac crescent, primitive heart tube, heart tube), as well as the integration of these samples together and exploration of that data. This data is provided as a general template for analysis rather than a line-by-line recreation of each figure in the paper, as many steps included in the analysis were repeated with different parameters (eg subclustering of the atrial cells vs subclustering of aSHF derivatives, etc). For any further questions, please feel free to email the authors at gonzalezdavidmatthew@gmail.com and we'll be happy to discuss further